Move specified and computed percentages to submodules
