+++ components/style/values/specified/rect.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-//! Computed types for CSS borders.
+//! Specified types for CSS borders.
 
 use cssparser::Parser;
 use parser::ParserContext;
